Change structure of files of tests for views
